Add files via upload
So far the coding is almost complete. 
The conversion from tabular data to graph data check.
Dividing dataset of a single graph with mask technique check.
The debugging of a correct graph neural network with graph convolutional layer check.
The debugging of the gradient descent, feedforward, and backpropagation check.
The code is fully functional.

There are still some minor problems with the policy used to determine the graph's edge.
It needs to be fixed. Otherwise the performance of the GNN will not be very ideal.